Start a kotlin based version of DB-WEAVE
This is the very first and entirely experimental
commit towards the new DB-WEAVE which will be based
on Kotlin.

The idea is to recreate DB-WEAVE from scratch using
modern technologies and languages. In a first step,
this means using Kotlin/JVM and Swing to build a
version for use on Desktop environments (i.e. Windows
and Linux). For macOS I probably will convert this
to Swift in order to enjoy all the apple stuff directly.
Mobile and web version are to be defined (android
probably derived from kotlin/swing version, iOS derived
from macOS/Swift version, web maybe using kotlin/js).