Refactor code and implement focus traversal
This extracts common code into the abstract
BaseView. Also, a custom focus traversal policy
is created and used to implement an easy way
to jump between the different components.

Still missing is a way to coordinate the current
position between the different components (e.g.
when jumping from pattern to threading, the
horizontal position should stay the same).