Manage CursorPos in addition to Selection
This allows to synchronize the cursor position
between the different components.

We still have some work to do, because now we
lose the selection, if focus goes away (e.g.
when alt-tabbing). This may be ok, but may be
annoying. A way to deal with this would be to
store one global selection for all components.