Use new Grid classes
This allows to remove a lot of repetitive model code.
Now, all components use the same Grid classes. This
makes code reuse much better. It may be slightly less
efficient, but in my tests it does not matter. What
does matter is a code as cleanly structured as possible.

The selection is now also shared by all components. Rather
surprisingly, this seems to work without further changes.
But I suspect, that we will have to add information about
the part the selection concerns. This will be important
when the selection becomes less ephemeral (i.e. is retained
when the focus is lost/acquired).