Simplify view code and separate concerns better
This removes special case views and instead introduces
GridView as a generic view. The callback object is
adjusted to call back abstract events. The controller
object then handles these events, changes the model
and repaints the view.

Right now, I am not at all content with the handling
of the different views, since the code is littered
with when statements that enumerate all views. I will
refactor this smell in one if the next commits. But
for now, I think that this commit is a step in the
right direction.